moved components to src/components folder
